// NordCode/NordCodeClient/src/app/models/menu.ts
export class Ecom_Menu {
  displayName: string;
  iconName: string;
  route: any;
  children: [];  
  ParentId: string;
  PID: any;

    // MenuId: any;
    // Title: string;
    // Path: string;
    // ParentId: any;
    // DisplayOrder: any;
    // Parameter: any;
    // NodeId: any;
    // Style: string;
    // Categories: [];
    // MenuRoles: string;
}